Dans le texte :
Click on Generate & Download BeerXML to downloard the .xml file. Put this file to the /public/recipes/your_beer_name/ directory will make your recipe visible in the My Recipes tab. Please note that the name of the .xml file must match the name of the directory it is placed in.

Example: a file named American-Stout.xml must be located in the /public/recipes/American-Stout/ directory on the github repository.

Passe une ligne entre Click on Generate & Download BeerXML to downloard the .xml file. Put this file to the /public/recipes/your_beer_name/ directory will make your recipe visible in the My Recipes tab.  et Please note that the name of the .xml file must match the name of the directory it is placed in.

De plus, seul la phrase Please note that the name of the .xml file must match the name of the directory it is placed in. doit être en gras et bleu